# Use optimized `append_..._n` methods in syslog receiver when constructing OTAP record