--- a/rust/otel-arrow-rust/src/encode/record/array/fixed_size_binary.rs
+++ b/rust/otel-arrow-rust/src/encode/record/array/fixed_size_binary.rs
@@ -193,11 +193,22 @@ mod test {
         CheckedArrayAppend::append_value(&mut fsb_builder, &b"1234".to_vec()).unwrap();
         CheckedArrayAppend::append_value(&mut fsb_builder, &b"5678".to_vec()).unwrap();
         CheckedArrayAppend::append_value(&mut fsb_builder, &b"9012".to_vec()).unwrap();
+        CheckedArrayAppendSlice::append_slice(&mut fsb_builder, b"4180").unwrap();
+        CheckedArrayAppendSlice::append_slice_n(&mut fsb_builder, b"5140", 2).unwrap();
+
         let result = ArrayBuilder::finish(&mut fsb_builder);
         assert_eq!(result.data_type(), &DataType::FixedSizeBinary(4));
 
         let expected = FixedSizeBinaryArray::try_from_iter(
-            [b"1234".to_vec(), b"5678".to_vec(), b"9012".to_vec()].iter(),
+            [
+                b"1234".to_vec(),
+                b"5678".to_vec(),
+                b"9012".to_vec(),
+                b"4180".to_vec(),
+                b"5140".to_vec(),
+                b"5140".to_vec(),
+            ]
+            .iter(),
         )
         .unwrap();
 
@@ -222,6 +233,11 @@ mod test {
         let index =
             CheckedDictionaryArrayAppend::append_value(&mut dict_builder, &b"b".to_vec()).unwrap();
         assert_eq!(index, 1);
+        let index = CheckedDictionaryAppendSlice::append_slice(&mut dict_builder, b"c").unwrap();
+        assert_eq!(index, 2);
+        let index =
+            CheckedDictionaryAppendSlice::append_slice_n(&mut dict_builder, b"d", 2).unwrap();
+        assert_eq!(index, 3);
 
         let result = DictionaryBuilder::finish(&mut dict_builder);
 
@@ -236,7 +252,10 @@ mod test {
         let mut expected_dict_values = FixedSizeBinaryBuilder::new(1);
         assert!(expected_dict_values.append_value(b"a").is_ok());
         assert!(expected_dict_values.append_value(b"b").is_ok());
-        let expected_dict_keys = UInt8Array::from_iter_values(vec![0, 0, 1]);
+        assert!(expected_dict_values.append_value(b"c").is_ok());
+        assert!(expected_dict_values.append_value(b"d").is_ok());
+        assert!(expected_dict_values.append_value(b"d").is_ok());
+        let expected_dict_keys = UInt8Array::from_iter_values(vec![0, 0, 1, 2, 3, 3]);
         let expected =
             UInt8DictionaryArray::new(expected_dict_keys, Arc::new(expected_dict_values.finish()));