Closes #879 

Use the "append_..._n" methods in the syslog receiver when constructing the arrow records. These methods have some nice performance optimizations:
- when appending a non-null value if the underlying builder is a dictionary array, the builder is able to append the value multiple times without doing multiple key lookups. For example:
https://github.com/apache/arrow-rs/blob/0737c61e76057b127312dd8058887649ece702b8/arrow-array/src/builder/generic_bytes_dictionary_builder.rs#L304-L316

- when the value is null, we can use the optimized code path to append multiple nulls to the null buffer:
https://github.com/apache/arrow-rs/blob/0737c61e76057b127312dd8058887649ece702b8/arrow-buffer/src/builder/boolean.rs#L193

Benchmark results:
```
arrow_batch_creation/rfc3164_arrow_batch_100_msgs
                        time:   [82.608 µs 82.774 µs 82.950 µs]
                        change: [−3.9462% −3.4305% −2.9199%] (p = 0.00 < 0.05)
                        Performance has improved.
                        
arrow_batch_creation/rfc5424_arrow_batch_100_msgs
                        time:   [46.139 µs 46.238 µs 46.338 µs]
                        change: [−4.2893% −3.9924% −3.7179%] (p = 0.00 < 0.05)
                        Performance has improved.
                        
arrow_batch_creation/cef_arrow_batch_100_msgs
                        time:   [40.973 µs 41.141 µs 41.308 µs]
                        change: [−3.8351% −3.3433% −2.8615%] (p = 0.00 < 0.05)
                        Performance has improved.
```